requirements.txt
boto3>=1.34.143
numpy>=1.24.3
pydub>=0.25.1
python-dotenv>=1.0.0
kokoro>=0.7.0
httpx>=0.24.1
torch>=2.0.1
transformers>=4.31.0
pymupdf>=1.25.0
soundfile
faiss-cpu>=1.11.0

torchaudio>=0.9.0
chatterbox-tts>=0.1.1